+++ ProjetFilmv1/AccueilPage.xaml.cs
@@ -19,47 +19,10 @@ namespace ProjetFilmv1
 
         private int _currentPage = 1;
         private int _totalPages = 1;
-        private int _currentGenreId = 0;
-
-        private bool _suppressPageComboChanged = false;
-
-        private Movie? _selectedMovie;
-        public Movie? SelectedMovie
-        {
-            get => _selectedMovie;
-            set
-            {
-                if (_selectedMovie != value)
-                {
-                    _selectedMovie = value;
-                    OnPropertyChanged(nameof(SelectedMovie));
-
-                    UpdatePaneLayout();
-
-                    if (_selectedMovie != null)
-                    {
-                        ShowDetailsInPane(_selectedMovie);
-                    }
-                    else
-                    {
-                        try
-                        {
-                            DetailsFrame?.NavigationService?.RemoveBackEntry();
-                            DetailsFrame.Content = null;
-                        }
-                        catch (Exception ex)
-                        {
-                            Debug.WriteLine($"Erreur remise à zéro DetailsFrame: {ex}");
-                        }
-                    }
-                }
-            }
-        }
+        private int _currentGenreId;
+        private bool _suppressPageComboChanged;
 
         public event PropertyChangedEventHandler? PropertyChanged;
-
-        protected void OnPropertyChanged(string name)
-            => PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(name));
 
         public AccueilPage()
         {
@@ -68,23 +31,30 @@ namespace ProjetFilmv1
             Loaded += AccueilPage_Loaded;
 
             for (int i = 1; i <= 20; i++)
+            {
                 PageComboBox.Items.Add(i);
+            }
 
             PageComboBox.SelectedIndex = 0;
 
-            PageComboBox.SelectionChanged += async (s, e) =>
+            PageComboBox.SelectionChanged += async (_, _) =>
             {
                 if (_suppressPageComboChanged)
+                {
                     return;
+                }
 
-            PageComboBox.SelectedIndex = 0;
-
-            GenreComboBox.SelectionChanged += async (s, e) =>
-            {
-                if (GenreComboBox.SelectedItem is Genre g)
-                    _currentGenreId = g.Id;
-                else
-                    _currentGenreId = 0;
+                if (PageComboBox.SelectedItem is int page)
+                {
+                    _currentPage = page;
+                    await LoadMoviesForCurrentSelectionAsync();
+                }
+                else if (PageComboBox.SelectedItem is object obj && int.TryParse(obj.ToString(), out var parsedPage))
+                {
+                    _currentPage = parsedPage;
+                    await LoadMoviesForCurrentSelectionAsync();
+                }
+            };
 
             GenreComboBox.SelectionChanged += async (_, _) =>
             {
@@ -102,61 +72,29 @@ namespace ProjetFilmv1
             {
                 if (_currentPage <= 1)
                 {
-                    _currentPage--;
-
-                    _suppressPageComboChanged = true;
-                    PageComboBox.SelectedItem = _currentPage;
-                    _suppressPageComboChanged = false;
-
-                    await LoadMoviesForCurrentSelectionAsync();
+                    return;
                 }
 
-            NextPageButton.Click += async (s, e) =>
-            {
-                if (_currentPage < _totalPages && _currentPage < 20)
-                {
-                    _currentPage++;
-
-                    _suppressPageComboChanged = true;
-                    PageComboBox.SelectedItem = _currentPage;
-                    _suppressPageComboChanged = false;
-
-                    await LoadMoviesForCurrentSelectionAsync();
-                }
+                _currentPage--;
+                _suppressPageComboChanged = true;
+                PageComboBox.SelectedItem = _currentPage;
+                _suppressPageComboChanged = false;
+                await LoadMoviesForCurrentSelectionAsync();
             };
 
-            LoadAllButton.Click += async (s, e) => await LoadAllPagesAsync();
-
-            for (int i = 1; i <= 20; i++)
+            NextPageButton.Click += async (_, _) =>
             {
-                var btn = new Button
+                if (_currentPage >= _totalPages || _currentPage >= 20)
                 {
-                    Content = i.ToString(),
-                    Width = 40,
-                    Margin = new Thickness(4),
-                    Tag = i
-                };
+                    return;
+                }
 
-                btn.Click += async (s, e) =>
-                {
-                    if (s is Button b && b.Tag is int page)
-                    {
-                        _currentPage = page;
-
-                        _suppressPageComboChanged = true;
-                        PageComboBox.SelectedItem = page;
-                        _suppressPageComboChanged = false;
-
-                        await LoadMoviesForCurrentSelectionAsync();
-                        UpdatePageButtons();
-                    }
-                };
-
-                PageButtonsPanel.Children.Add(btn);
-            }
-
-            UpdatePageButtons();
-            UpdatePaneLayout();
+                _currentPage++;
+                _suppressPageComboChanged = true;
+                PageComboBox.SelectedItem = _currentPage;
+                _suppressPageComboChanged = false;
+                await LoadMoviesForCurrentSelectionAsync();
+            };
         }
 
         private async void AccueilPage_Loaded(object sender, RoutedEventArgs e)
@@ -170,13 +108,13 @@ namespace ProjetFilmv1
             try
             {
                 var list = await _tmdb.GetGenresAsync();
-                Debug.WriteLine($"Genres loaded: {list.Count}");
-
                 Genres.Clear();
                 Genres.Add(new Genre { Id = 0, Name = "Tous" });
 
-                foreach (var g in list)
-                    Genres.Add(g);
+                foreach (var genre in list)
+                {
+                    Genres.Add(genre);
+                }
 
                 GenreComboBox.ItemsSource = Genres;
                 GenreComboBox.SelectedIndex = 0;
@@ -196,21 +134,15 @@ namespace ProjetFilmv1
         {
             try
             {
-                Debug.WriteLine($"Loading movies for genre {_currentGenreId} page {_currentPage}...");
-
-                TmdbResponse? resp;
-                if (_currentGenreId == 0)
-                    resp = await _tmdb.GetPopularMoviesResponseAsync(_currentPage);
-                else
-                    resp = await _tmdb.GetMoviesByGenreResponseAsync(_currentGenreId, _currentPage);
-
-                var movies = resp?.Results ?? new System.Collections.Generic.List<Movie>();
-
-                Debug.WriteLine($"Movies loaded: {movies.Count}");
+                TmdbResponse? response = _currentGenreId == 0
+                    ? await _tmdb.GetPopularMoviesResponseAsync(_currentPage)
+                    : await _tmdb.GetMoviesByGenreResponseAsync(_currentGenreId, _currentPage);
 
                 Movies.Clear();
-                foreach (var m in movies)
-                    Movies.Add(m);
+                foreach (var movie in response?.Results ?? [])
+                {
+                    Movies.Add(movie);
+                }
 
                 _totalPages = response?.TotalPages ?? 1;
                 PageText.Text = $"Page {_currentPage} / {_totalPages}";
@@ -218,8 +150,7 @@ namespace ProjetFilmv1
                 try
                 {
                     _suppressPageComboChanged = true;
-
-                    if (PageComboBox.Items.Count >= 1)
+                    if (PageComboBox.Items.Count > 0)
                     {
                         PageComboBox.SelectedIndex = Math.Max(0, Math.Min(PageComboBox.Items.Count - 1, _currentPage - 1));
                     }
@@ -233,16 +164,6 @@ namespace ProjetFilmv1
 
                 PrevPageButton.IsEnabled = _currentPage > 1;
                 NextPageButton.IsEnabled = _currentPage < _totalPages && _currentPage < 20;
-
-                if (PageButtonsPanel != null)
-                {
-                    PageButtonsPanel.Visibility = Visibility.Visible;
-                    UpdatePageButtons();
-                }
-
-                MoviesListBox.SelectedItem = null;
-                SelectedMovie = null;
-                UpdatePaneLayout();
             }
             catch (Exception ex)
             {
@@ -255,146 +176,35 @@ namespace ProjetFilmv1
             }
         }
 
-        private async Task ReloadCurrentAsync()
+        private void OnMovieButtonClick(object sender, RoutedEventArgs e)
         {
-            await LoadMoviesForCurrentSelectionAsync();
-        }
+            if (sender is not Button button || button.DataContext is not Movie movie)
+            {
+                return;
+            }
 
-        private void ShowDetailsInPane(Movie movie)
-        {
             try
             {
-                if (movie == null)
-                    return;
+                var owner = Window.GetWindow(this);
+                var detailsWindow = new MovieDetailsWindow(movie)
+                {
+                    Owner = owner,
+                    WindowStartupLocation = WindowStartupLocation.CenterOwner
+                };
 
-                var detailsPage = new MovieDetailsWindow2(movie);
-                DetailsFrame.Navigate(detailsPage);
+                detailsWindow.ShowDialog();
             }
             catch (Exception ex)
             {
-                Debug.WriteLine($"Erreur ShowDetailsInPane: {ex}");
-                MessageBox.Show(
-                    $"Erreur en affichant le détail : {ex.Message}",
-                    "Erreur",
-                    MessageBoxButton.OK,
-                    MessageBoxImage.Error);
+                Debug.WriteLine($"Erreur ouverture détail film: {ex}");
+                MessageBox.Show($"Erreur en affichant le détail : {ex.Message}", "Erreur", MessageBoxButton.OK, MessageBoxImage.Error);
             }
         }
 
-        private async Task LoadAllPagesAsync()
+        protected void OnPropertyChanged(string propertyName)
         {
-            try
-            {
-                _currentPage = 1;
-                Movies.Clear();
-
-                TmdbResponse? firstResp;
-                if (_currentGenreId == 0)
-                    firstResp = await _tmdb.GetPopularMoviesResponseAsync(1);
-                else
-                    firstResp = await _tmdb.GetMoviesByGenreResponseAsync(_currentGenreId, 1);
-
-                if (firstResp == null)
-                    return;
-
-                var total = firstResp.TotalPages;
-                Debug.WriteLine($"LoadAll: total pages = {total}");
-
-                foreach (var m in firstResp.Results)
-                    Movies.Add(m);
-
-                for (int p = 2; p <= total; p++)
-                {
-                    await Task.Delay(250);
-
-                    TmdbResponse? resp;
-                    if (_currentGenreId == 0)
-                        resp = await _tmdb.GetPopularMoviesResponseAsync(p);
-                    else
-                        resp = await _tmdb.GetMoviesByGenreResponseAsync(_currentGenreId, p);
-
-                    if (resp == null)
-                        break;
-
-                    foreach (var m in resp.Results)
-                        Movies.Add(m);
-                }
-
-                _currentPage = 1;
-                _totalPages = total;
-
-                _suppressPageComboChanged = true;
-                PageComboBox.SelectedIndex = 0;
-                _suppressPageComboChanged = false;
-
-                PageText.Text = $"Page {_currentPage} / {_totalPages} (chargé)";
-                PrevPageButton.IsEnabled = false;
-                NextPageButton.IsEnabled = false;
-
-                MoviesListBox.SelectedItem = null;
-                SelectedMovie = null;
-                UpdatePaneLayout();
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine($"Erreur LoadAllPagesAsync: {ex}");
-                MessageBox.Show(
-                    $"Erreur lors du chargement total : {ex.Message}",
-                    "Erreur",
-                    MessageBoxButton.OK,
-                    MessageBoxImage.Error);
-            }
-        }
-
-        private void UpdatePageButtons()
-        {
-            foreach (var child in PageButtonsPanel.Children)
-            {
-                if (child is Button btn && btn.Tag is int page)
-                {
-                    btn.IsEnabled = true;
-                    btn.Opacity = (page == _currentPage) ? 1.0 : 0.5;
-                }
-            }
-        }
-
-        private void UpdatePaneLayout()
-        {
-            try
-            {
-                if (LeftColumn == null || RightColumn == null)
-                    return;
-
-                if (SelectedMovie != null)
-                {
-                    LeftColumn.Width = new GridLength(1, GridUnitType.Star);
-                    RightColumn.Width = new GridLength(2, GridUnitType.Star);
-                }
-                else
-                {
-                    LeftColumn.Width = new GridLength(1, GridUnitType.Star);
-                    RightColumn.Width = new GridLength(1, GridUnitType.Star);
-                }
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine($"Erreur UpdatePaneLayout: {ex}");
-            }
-        }
-
-        private void MoviesListBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
-        {
-            try
-            {
-                if (MoviesListBox.SelectedItem is Movie movie)
-                    SelectedMovie = movie;
-                else
-                    SelectedMovie = null;
-            }
-            catch (Exception ex)
-            {
-                Debug.WriteLine($"Erreur MoviesListBox_SelectionChanged: {ex}");
-            }
+            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
         }
     }
 }
+
