+++ ProjetFilmv1/MovieDetailsWindow.xaml
@@ -59,8 +59,17 @@
                                 <RowDefinition Height="Auto"/>
                             </Grid.RowDefinitions>
 
-                            <TextBlock Text="Commentaires" FontSize="18" FontWeight="SemiBold" Margin="0,0,0,10" Foreground="{DynamicResource TextBrush}"/>
-                            <ScrollViewer Grid.Row="1" VerticalScrollBarVisibility="Auto">
+                            <TextBlock 
+                                Text="Commentaires" 
+                                FontSize="18" 
+                                FontWeight="SemiBold" 
+                                Margin="0,0,0,10" 
+                                Foreground="{DynamicResource TextBrush}"/>
+                            
+
+                            <ScrollViewer Height="200"
+                                          VerticalScrollBarVisibility="Auto"
+                                          Margin="0,0,0,10">
                                 <StackPanel x:Name="CommentsPanel"/>
                             </ScrollViewer>
                             <TextBox x:Name="CommentTextBox"
